@@ -19,21 +19,49 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.2.0 <4.0.0"
+  sdk: ">=3.4.3 <4.0.0"
 
 dependencies:
+  auto_route: ^8.2.0
   cupertino_icons: ^1.0.2
+  dio: ^5.5.0
   equatable: ^2.0.5
+  firebase_auth: ^5.1.0
+  firebase_core: ^3.1.1
+  firebase_storage: ^12.1.0
   flutter:
     sdk: flutter
+  flutter_bloc: ^8.1.6
+  flutter_localizations:
+    sdk: flutter
+  flutter_secure_storage: ^9.2.2
+  flutter_form_builder: ^9.3.0
+  form_builder_validators: ^10.0.1
+  font_awesome_flutter: ^10.7.0
+  sign_button: ^2.0.6
+  google_sign_in: ^6.2.1
+  intl: ^0.19.0
+  provider: ^6.1.2
+  retrofit: ^4.1.0
+  talker: ^4.3.2
+  talker_bloc_logger: ^4.3.2
+  window_manager: ^0.3.9
 
 dev_dependencies:
+  auto_route_generator: ^8.0.0
+  bloc_test: ^9.1.7
+  build_runner: ^2.4.11
+  data_fixture_dart: ^2.2.0
   flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
+  http_mock_adapter: ^0.6.1
+  mockito: ^5.4.4
+  retrofit_generator: ^8.1.1
 
 flutter:
   uses-material-design: true
+  generate: true
 
   assets:
     - assets/images/
