@@ -1,6 +1,46 @@
-import 'package:flutter/material.dart';
-import 'package:twelve_notes/src/app.dart';
+import 'dart:developer';
+import 'dart:io';
 
-void main() {
-  runApp(const App());
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:talker/talker.dart';
+import 'package:talker_bloc_logger/talker_bloc_logger_observer.dart';
+import 'package:twelve_notes/firebase_options.dart';
+import 'package:twelve_notes/src/app.dart';
+import 'package:window_manager/window_manager.dart';
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+
+  await FirebaseAuth.instance.useAuthEmulator('127.0.0.1', 9099);
+
+  Bloc.observer = TalkerBlocObserver(
+    talker: Talker(
+      logger: TalkerLogger(
+        output: log,
+      ),
+    ),
+  );
+
+  SystemChrome.setPreferredOrientations([
+    DeviceOrientation.portraitUp,
+  ]);
+
+  if (Platform.isLinux || Platform.isMacOS || Platform.isWindows) {
+    await windowManager.ensureInitialized();
+
+    await windowManager.setMinimumSize(
+      const Size(300.0, 600.0),
+    );
+  }
+  runApp(
+    const App(),
+  );
 }
